Rename new gazer file to remove v2 suffix
