+++ pupil_src/shared_modules/gaze_mapping/__init__.py
@@ -12,7 +12,7 @@ from . import matching
 from . import gazer_base
 from .gazer_base import CalibrationError
 from .gazer_2d_v1x import Gazer2D_v1x
-from .gazer_2d_v2 import Gazer2D
+from .gazer_2d import Gazer2D
 from .gazer_3d_v1x import Gazer3D_v1x
 from .gazer_hmd_v1x import GazerHMD3D_v1x
